DependencyCheck and CircularDep improvements
* include/CircularDep.hpp
Super alpha version, a lot of mess in code, but it works.

* include/DependencyCheck.hpp
Added loading dependencies to graph and processing them.

* include/filedesc_graph.hpp
Namespace fix.

We're heading toward a lot of changes and files renaming, because
DependencyCheck and derived classes are going to be rewritten to
implement decorator design pattern -- so get ready.